move activity level and surplus to macros
