+++ src/lib/components/forms/Macros.svelte
@@ -4,10 +4,20 @@
 	import { determinePRatio } from '../../../utils/PRatioDeterminer';
 	import { calculateFatMass, calculateLeanMass } from '../../../utils/MassCalculator';
 	import { calculateMacronutrients } from '../../../utils/NutritionCalculations';
+	import { RadioGroup, RadioItem } from '@skeletonlabs/skeleton';
 
 	let bodyFat = 10;
 	let bodyFatMin = 4;
 	let bodyFatMax = 20;
+
+	const activityLevels = {
+		sedentary: 'Sedentary',
+		lightlyActive: 'Light',
+		moderatelyActive: 'Moderate',
+		veryActive: 'High',
+		extraActive: 'Extreme'
+	};
+	const calorieSurplusRange = [2.5, 5, 10, 15, 25];
 
 	$: if ($UserStore.sex === 'male') {
 		bodyFat = 10;
@@ -68,7 +78,23 @@
 	}
 </script>
 
-<section class="px-4 pt-8 pb-28">
+<section class="px-4 pb-8">
+	<form class="flex flex-col justify-between gap-5 px-2 mt-6 mb-8 max-w-xl m-auto">
+		<p class="unstyled text-lg">Increasing calorie intake by:</p>
+		<RadioGroup class="flex items-center">
+			{#each calorieSurplusRange as value}
+				<RadioItem type="radio" {value} bind:group={$UserStore.calorieSurplus} name="" id=""
+					>{value}%</RadioItem
+				>
+			{/each}
+		</RadioGroup>
+		<p class="font-semibold">Select your activity level:</p>
+		<RadioGroup display="flex-col" class="w-1/2 m-auto">
+			{#each Object.entries(activityLevels) as [value, name]}
+				<RadioItem {value} bind:group={$UserStore.activityLevel} {name}>{name}</RadioItem>
+			{/each}
+		</RadioGroup>
+	</form>
 	{#if errorMessage}
 		<p class="unstyled text-xl">{errorMessage}</p>
 	{:else}
